WIP - remove "/esc" from base path
All of the SDK API endpoints are prefixed with /api/esc so we need to
rework that to get change-requests submit going (both for
/api/preview/esc and /api/preview/change-requests). It will be a little
odd to have core Pulumi change request submit functionality in the ESC
SDK but it is necessary. We let users specify the "host" today in the
SDK and it has to include /api/esc so adjusting that to support preview
URLs and higher level URLs is going to be a little tricky.

This commit partially works toward stripping out /api/esc from the user
provided host in order to stay backwards compatible. See TODO and this
is completely untested and not complete. I was moved off of this task
after starting, recording my progress.